Fix dummies
New strategy for handling dummies:

Start by canonicalizing the contractions to a predefined form, in this
case choosen to be a series of (Contract 0 1).

This can always be done since a general contraction (Contract a b expr)
can be replaced by a combination of (Contract 0 1 (Permute [...] expr)
where the permutation is such that 0 -> a and 1 -> b.

This permutation is found by taking the range [3..] and inserting
1 & 2 at the positions a and b.

After this every monoterm consists of a series of contractions and a
permutation
(C 0 1 (C 0 1 ... (Permute p expr)))
The permutation can now be canonicalized as free indices keeping the
contractions fixed. Since all terms with the same number of contractions
will map to exactly the same contraction structure this completely
canonicalizes dummy indices.